give application’s files their own directory
