+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "NODE_ENV=dev node server.js",
     "test": "NODE_ENV=test tape test/*.js | tap-spec",
-    "travis-test": "NODE_ENV=test node ./test/*.js --cov | tap-spec"
+    "travis-test": "NODE_ENV=test tap test/*.js | tap-spec"
   },
   "repository": {
     "type": "git",
